requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-PyMongo==2.3.0
greenlet==1.0.0
gunicorn==20.1.0
idna==2.10
importlib-metadata==3.8.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pymongo==3.11.3
pyTelegramBotAPI==3.7.7
python-dotenv==0.16.0
requests==2.25.1
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
zipp==3.4.1
